Refactor network list
