+++ cmd/nerdctl/network_ls.go
@@ -17,15 +17,8 @@
 package main
 
 import (
-	"bytes"
-	"errors"
-	"fmt"
-	"text/tabwriter"
-	"text/template"
-
-	"github.com/containerd/nerdctl/pkg/formatter"
-	"github.com/containerd/nerdctl/pkg/netutil"
-
+	"github.com/containerd/nerdctl/pkg/api/types"
+	"github.com/containerd/nerdctl/pkg/cmd/network"
 	"github.com/spf13/cobra"
 )
 
@@ -48,14 +41,6 @@ func newNetworkLsCommand() *cobra.Command {
 	return cmd
 }
 
-type networkPrintable struct {
-	ID     string // empty for non-nerdctl networks
-	Name   string
-	Labels string
-	// TODO: "CreatedAt", "Driver", "IPv6", "Internal", "Scope"
-	file string `json:"-"`
-}
-
 func networkLsAction(cmd *cobra.Command, args []string) error {
 	globalOptions, err := processRootCmdFlags(cmd)
 	if err != nil {
@@ -65,86 +50,14 @@ func networkLsAction(cmd *cobra.Command, args []string) error {
 	if err != nil {
 		return err
 	}
-	w := cmd.OutOrStdout()
-	var tmpl *template.Template
 	format, err := cmd.Flags().GetString("format")
 	if err != nil {
 		return err
 	}
-	switch format {
-	case "", "table", "wide":
-		w = tabwriter.NewWriter(cmd.OutOrStdout(), 4, 8, 4, ' ', 0)
-		if !quiet {
-			fmt.Fprintln(w, "NETWORK ID\tNAME\tFILE")
-		}
-	case "raw":
-		return errors.New("unsupported format: \"raw\"")
-	default:
-		if quiet {
-			return errors.New("format and quiet must not be specified together")
-		}
-		var err error
-		tmpl, err = formatter.ParseTemplate(format)
-		if err != nil {
-			return err
-		}
+	options := types.NetworkListCommandOptions{
+		GOptions: globalOptions,
+		Quiet:    quiet,
+		Format:   format,
 	}
-
-	e, err := netutil.NewCNIEnv(globalOptions.CNIPath, globalOptions.CNINetConfPath)
-	if err != nil {
-		return err
-	}
-	netConfigs, err := e.NetworkList()
-	if err != nil {
-		return err
-	}
-	pp := make([]networkPrintable, len(netConfigs))
-	for i, n := range netConfigs {
-		p := networkPrintable{
-			Name: n.Name,
-			file: n.File,
-		}
-		if n.NerdctlID != nil {
-			p.ID = *n.NerdctlID
-			if len(p.ID) > 12 {
-				p.ID = p.ID[:12]
-			}
-		}
-		if n.NerdctlLabels != nil {
-			p.Labels = formatter.FormatLabels(*n.NerdctlLabels)
-		}
-		pp[i] = p
-	}
-
-	// append pseudo networks
-	pp = append(pp, []networkPrintable{
-		{
-			Name: "host",
-		},
-		{
-			Name: "none",
-		},
-	}...)
-
-	for _, p := range pp {
-		if tmpl != nil {
-			var b bytes.Buffer
-			if err := tmpl.Execute(&b, p); err != nil {
-				return err
-			}
-			if _, err = fmt.Fprintf(w, b.String()+"\n"); err != nil {
-				return err
-			}
-		} else if quiet {
-			if p.ID != "" {
-				fmt.Fprintln(w, p.ID)
-			}
-		} else {
-			fmt.Fprintf(w, "%s\t%s\t%s\n", p.ID, p.Name, p.file)
-		}
-	}
-	if f, ok := w.(formatter.Flusher); ok {
-		return f.Flush()
-	}
-	return nil
+	return network.List(cmd.Context(), options, cmd.OutOrStdout())
 }
